requirements.txt
asgiref==3.6.0
Django==4.2
django-ajax-datatable==4.4.5
sqlparse==0.4.4
